casting for output

+++ CQuiz/test.c
@@ -7,5 +7,5 @@ int main(){
 	int B = -1; 
 	int grade[5] = { 4, 3, 2, 1, 0 };
 
-	printf("%d, %d\n", strlen(cp), sizeof(grade));
+	printf("%d, %d\n", (int)strlen(cp),(int) sizeof(grade));
 }
